refactor(meta): reorganize code of global barrier manager (part 1)
